PRSD-372: Renames SummaryPages to CheckAnswers pages
