+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/forms/journeys/LaUserRegistrationJourney.kt
@@ -3,7 +3,7 @@ package uk.gov.communities.prsdb.webapp.forms.journeys
 import org.springframework.stereotype.Component
 import org.springframework.validation.Validator
 import uk.gov.communities.prsdb.webapp.constants.enums.JourneyType
-import uk.gov.communities.prsdb.webapp.forms.pages.LaUserRegistrationSummaryPage
+import uk.gov.communities.prsdb.webapp.forms.pages.LaUserRegistrationCheckAnswersPage
 import uk.gov.communities.prsdb.webapp.forms.pages.Page
 import uk.gov.communities.prsdb.webapp.forms.steps.RegisterLaUserStepId
 import uk.gov.communities.prsdb.webapp.forms.steps.Step
@@ -93,7 +93,7 @@ class LaUserRegistrationJourney(
         ) = Step(
             id = RegisterLaUserStepId.CheckAnswers,
             page =
-                LaUserRegistrationSummaryPage(
+                LaUserRegistrationCheckAnswersPage(
                     formModel = CheckAnswersFormModel::class,
                     templateName = "forms/checkAnswersForm",
                     content =
