+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/forms/pages/LaUserRegistrationCheckAnswersPage.kt
@@ -3,45 +3,60 @@ package uk.gov.communities.prsdb.webapp.forms.pages
 import org.springframework.ui.Model
 import org.springframework.validation.Validator
 import uk.gov.communities.prsdb.webapp.constants.enums.JourneyType
-import uk.gov.communities.prsdb.webapp.forms.journeys.JourneyData
+import uk.gov.communities.prsdb.webapp.exceptions.PrsdbWebException
 import uk.gov.communities.prsdb.webapp.forms.journeys.objectToStringKeyedMap
 import uk.gov.communities.prsdb.webapp.forms.steps.RegisterLaUserStepId
 import uk.gov.communities.prsdb.webapp.models.dataModels.FormSummaryDataModel
 import uk.gov.communities.prsdb.webapp.models.formModels.FormModel
+import uk.gov.communities.prsdb.webapp.services.JourneyDataService
+import uk.gov.communities.prsdb.webapp.services.LocalAuthorityInvitationService
 import kotlin.reflect.KClass
 
-class LandlordRegistrationSummaryPage(
+class LaUserRegistrationCheckAnswersPage(
     formModel: KClass<out FormModel>,
     templateName: String,
     content: Map<String, Any>,
+    private val journeyDataService: JourneyDataService,
+    private val invitationService: LocalAuthorityInvitationService,
 ) : Page(formModel, templateName, content) {
     override fun populateModelAndGetTemplateName(
         validator: Validator,
         model: Model,
         pageData: Map<String, Any?>?,
         prevStepUrl: String?,
-        journeyData: JourneyData?,
     ): String {
+        val journeyData = journeyDataService.getJourneyDataFromSession()
         val formData = mutableListOf<FormSummaryDataModel>()
+        val sessionToken = invitationService.getTokenFromSession()
 
-        // TODO PRSD-372 update the formData below
+        val localAuthority =
+            if (sessionToken != null) {
+                invitationService.getAuthorityForToken(sessionToken)
+            } else {
+                throw PrsdbWebException("Local authority not found for this invitation token")
+            }
 
         formData.addAll(
             listOf(
                 FormSummaryDataModel(
+                    "registerLaUser.checkAnswers.rowHeading.localAuthority",
+                    localAuthority?.name,
+                    null,
+                ),
+                FormSummaryDataModel(
                     "registerLaUser.checkAnswers.rowHeading.name",
-                    objectToStringKeyedMap(journeyData?.get("name"))?.get("name"),
+                    objectToStringKeyedMap(journeyData["name"])?.get("name"),
                     "/${JourneyType.LA_USER_REGISTRATION.urlPathSegment}/${RegisterLaUserStepId.Name.urlPathSegment}",
                 ),
                 FormSummaryDataModel(
                     "registerLaUser.checkAnswers.rowHeading.email",
-                    objectToStringKeyedMap(journeyData?.get("email"))?.get("emailAddress"),
+                    objectToStringKeyedMap(journeyData["email"])?.get("emailAddress"),
                     "/${JourneyType.LA_USER_REGISTRATION.urlPathSegment}/${RegisterLaUserStepId.Email.urlPathSegment}",
                 ),
             ),
         )
 
         model.addAttribute("formData", formData)
-        return super.populateModelAndGetTemplateName(validator, model, pageData, prevStepUrl, journeyData)
+        return super.populateModelAndGetTemplateName(validator, model, pageData, prevStepUrl)
     }
 }
